updated reading and filtering of search results:
- one static filtering function for all formats
- also filter out out "O" = Pyrrolysin
