+++ spectrum_io/search_result/msfragger.py
@@ -4,7 +4,7 @@ import pandas as pd
 import spectrum_fundamentals.constants as c
 from spectrum_fundamentals.mod_string import internal_without_mods
 
-from .search_results import SearchResults
+from .search_results import SearchResults, filter_valid_prosit_sequences
 
 logger = logging.getLogger(__name__)
 
@@ -52,6 +52,18 @@ class MSFragger(SearchResults):
         df.columns = df.columns.str.upper()
         df.columns = df.columns.str.replace(" ", "_")
 
+        df = MSFragger.update_columns_for_prosit(df, tmt_labeled)
+        return filter_valid_prosit_sequences(df)
+
+    @staticmethod
+    def update_columns_for_prosit(df, tmt_labeled: str) -> pd.DataFrame:
+        """
+        Update columns of df to work with Prosit.
+
+        :param df: df to modify
+        :param tmt_labeled: True if tmt labeled
+        :return: modified df as pd.DataFrame
+        """
         df.rename(columns={"CHARGE": "PRECURSOR_CHARGE"}, inplace=True)
 
         df["REVERSE"] = df["PROTEIN"].str.contains("Reverse")
@@ -83,12 +95,4 @@ class MSFragger(SearchResults):
         df["SEQUENCE"] = internal_without_mods(df["MODIFIED_SEQUENCE"])
         df["PEPTIDE_LENGTH"] = df["SEQUENCE"].apply(lambda x: len(x))
 
-        logger.info(f"No of sequences before Filtering is {len(df['PEPTIDE_LENGTH'])}")
-        df = df[(df["PEPTIDE_LENGTH"] <= 30)]
-        df = df[(~df["MODIFIED_SEQUENCE"].str.contains(r"\(ac\)"))]
-        df = df[(~df["MODIFIED_SEQUENCE"].str.contains(r"\(Acetyl \(Protein N-term\)\)"))]
-        df = df[(~df["SEQUENCE"].str.contains("U"))]
-        df = df[df["PRECURSOR_CHARGE"] <= 6]
-        df = df[df["PEPTIDE_LENGTH"] >= 7]
-        logger.info(f"No of sequences after Filtering is {len(df['PEPTIDE_LENGTH'])}")
         return df
